working using second swagger
